backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
redis>=5.0.0
celery>=5.3.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0
aiohttp>=3.9.0
websockets>=12.0
aiofiles>=23.2.0
rich>=13.7.0
loguru>=0.7.0
tenacity>=8.2.0
pydantic-settings>=2.1.0
python-nmap>=0.7.1
requests>=2.31.0
beautifulsoup4>=4.12.0
dnspython>=2.4.0
python-whois>=0.8.0
jinja2>=3.1.0
markdown>=3.5.0
weasyprint>=61.0
pdfkit>=1.0.0
reportlab>=4.0.0
playwright>=1.40.0
google-generativeai>=0.3.0
aiosqlite>=0.21.0
# Additional security tools dependencies
beautifulsoup4>=4.12.0  # Already included but ensure version
lxml>=4.9.3  # For XML parsing (XXE agent)
cssselect>=1.2.0  # For advanced web parsing
tld>=0.13  # For domain parsing
validators>=0.20.0  # For URL validation
python-dateutil>=2.8.2  # For date parsing
chardet>=5.2.0  # For character encoding detection
# Network and security scanning
scapy>=2.5.0  # For network packet manipulation
netaddr>=0.8.0  # For network address manipulation
ipaddress>=1.0.0  # For IP address handling
# SSL/TLS testing
pyopenssl>=23.2.0  # For SSL/TLS certificate handling
cryptography>=41.0.0  # For cryptographic operations
# Web security testing
selenium>=4.15.0  # For browser automation (DOM XSS)
requests-html>=0.10.0  # For JavaScript rendering
fake-useragent>=1.4.0  # For user agent rotation
# API testing
jsonschema>=4.19.0  # For API schema validation
pyjwt>=2.8.0  # For JWT token handling
urllib3>=2.0.0  # For advanced HTTP handling
# Performance and async
aiofiles>=23.2.0  # Already included
asyncio-throttle>=1.0.2  # For request throttling
# Database support
pymongo>=4.5.0  # For NoSQL injection testing
redis>=5.0.0  # Already included
# Additional parsing and analysis
python-magic>=0.4.27  # For file type detection
pillow>=10.0.0  # For image processing
pypdf2>=3.0.0  # For PDF analysis
# Logging and monitoring
structlog>=23.1.0  # For structured logging
# Additional dependencies for new security agents
zipfile36>=0.1.3  # For zip file handling (file upload agent)
python-magic>=0.4.27  # File type detection
pyjwt[crypto]>=2.8.0  # For JWT security testing
bcrypt>=4.1.0  # For password hashing tests
cffi>=1.16.0  # For cryptographic operations
# WebSocket and real-time testing
websockets>=12.0  # For WebSocket security testing
websocket-client>=1.6.0  # WebSocket client library
# LDAP testing dependencies
python-ldap>=3.4.0  # LDAP client library (may require system libraries)
ldap3>=2.9.0  # Pure Python LDAP client (alternative)
# Business logic and advanced testing
parameterized>=0.8.1  # For parameterized testing
hypothesis>=6.0.0  # Property-based testing for edge cases
# Supply Chain Security dependencies
packaging>=23.0  # For package version parsing and analysis
toml>=0.10.2  # For parsing pyproject.toml and config files
semantic-version>=2.10.0  # For dependency version analysis
# AI/ML Security testing dependencies
transformers>=4.30.0  # For AI model interaction (optional)
requests-oauthlib>=1.3.0  # For API authentication testing
jsonlines>=4.0.0  # For processing JSONL data formats
# Advanced security analysis
python-whois>=0.8.0  # For domain analysis
dnspython>=2.4.0  # For advanced DNS analysis
certifi>=2023.0.0  # For certificate validation 